+++ src/features/partyroom/change-my-playlist/index.ts
@@ -1,0 +1,1 @@
+export { useChangeMyPlaylist } from './api/use-change-my-playlist.mutation';
